Added ppi feature
Added known ppi values for applicable devices.

Unit test added and verified with the devices public information in WikiPedia.

Applicable devices are all devices that include a screen, so basically all the iPhones and Ipads.

Unaplicable devices are devices whose screen is not part of the device itself, such as the AppleTV and the Simulators.