Fix missing os warning in Swift Playgrounds
In Swift Playgrounds (which doesn't yet support Swift 5.9), there is a warning that the OS is not supported since it doesn't know about xrOS (visionOS): "Unknown operating system for build configuration 'os'".  I've re-worked the conditional compiles to look for the module UIScreen which is more appropriate rather than the OS tests and added a Swift version compiler check for the case that isn't module dependent.